Allow the option of synchronous input handling.
This is a bit of a twitchy tradeoff:  Slow handlers would be a serious UX issue.  On the other hand, creating a thread for every input is insane and (apparently) one of my biggest sources of allocations right now.

So, to support long-running processes more gracefully, where GC pressure can be a problem, I introduce this.